refactor: Move `OpBox` to a new module
