+++ src/lib.rs
@@ -3,9 +3,13 @@
 //! [TKET](https://github.com/CQCL/tket) quantum compiler.
 
 pub mod circuit_json;
+pub mod opbox;
 pub mod optype;
 #[cfg(feature = "pyo3")]
 pub mod pytket;
+
+pub use circuit_json::SerialCircuit;
+pub use optype::OpType;
 
 #[cfg(test)]
 mod tests {
